rename lipid packages
