+++ mzmine-community/src/main/java/io/github/mzmine/modules/dataprocessing/id_lipididentification/common/lipididentification/matches/species/FattyAcylSpeciesLevelMatchedLipidFactory.java
@@ -23,26 +23,10 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package io.github.mzmine.modules.dataprocessing.id_lipididentification.common.lipididentificationtools;
+package io.github.mzmine.modules.dataprocessing.id_lipididentification.common.lipididentification.matches.species;
 
-public enum LipidFragmentationRuleRating {
-  MINOR("minor", "Fragmentation rule alone is not enough to identify a lipid"),//
-  MAJOR("major", "Fragmentation rule can identify lipid");//
+public class FattyAcylSpeciesLevelMatchedLipidFactory extends
+    GlyceroAndGlycerophosphoSpeciesLevelMatchedLipidFactory implements
+    ISpeciesLevelMatchedLipidFactory {
 
-  private final String name;
-  private final String description;
-
-  LipidFragmentationRuleRating(String name, String description) {
-    this.name = name;
-    this.description = description;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  public String getDescription() {
-    return description;
-  }
 }
-
