+++ mzmine-community/src/main/java/io/github/mzmine/modules/dataprocessing/id_lipididentification/common/lipididentification/matches/species/ISpeciesLevelMatchedLipidFactory.java
@@ -23,10 +23,21 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package io.github.mzmine.modules.dataprocessing.id_lipididentification.common.lipididentificationtools.matchedlipidannotations.specieslevellipidmatches;
+package io.github.mzmine.modules.dataprocessing.id_lipididentification.common.lipididentification.matches.species;
 
-public class FattyAcylSpeciesLevelMatchedLipidFactory extends
-    GlyceroAndGlycerophosphoSpeciesLevelMatchedLipidFactory implements
-    ISpeciesLevelMatchedLipidFactory {
+import io.github.mzmine.datamodel.DataPoint;
+import io.github.mzmine.datamodel.IonizationType;
+import io.github.mzmine.modules.dataprocessing.id_lipididentification.common.lipididentification.matches.MatchedLipid;
+import io.github.mzmine.modules.dataprocessing.id_lipididentification.common.lipids.ILipidAnnotation;
+import io.github.mzmine.modules.dataprocessing.id_lipididentification.common.lipids.LipidFragment;
+import io.github.mzmine.parameters.parametertypes.tolerances.MZTolerance;
+import java.util.Set;
+
+public interface ISpeciesLevelMatchedLipidFactory {
+
+  MatchedLipid validateSpeciesLevelAnnotation(double accurateMz,
+      ILipidAnnotation speciesLevelAnnotation, Set<LipidFragment> annotatedFragments,
+      DataPoint[] massList, double minMsMsScore, MZTolerance mzTolRangeMSMS,
+      IonizationType ionizationType);
 
 }
