+++ mzmine-community/src/main/java/io/github/mzmine/modules/visualization/spectra/matchedlipid/LipidAnnotationMatchPaneOld.java
@@ -27,7 +27,7 @@ package io.github.mzmine.modules.visualization.spectra.matchedlipid;
 
 import io.github.mzmine.gui.chartbasics.simplechart.datasets.RunOption;
 import io.github.mzmine.main.MZmineCore;
-import io.github.mzmine.modules.dataprocessing.id_lipididentification.common.lipididentificationtools.matchedlipidannotations.MatchedLipid;
+import io.github.mzmine.modules.dataprocessing.id_lipididentification.common.lipididentification.matches.MatchedLipid;
 import io.github.mzmine.modules.dataprocessing.id_lipididentification.common.lipids.LipidFragment;
 import io.github.mzmine.util.FormulaUtils;
 import io.github.mzmine.util.color.ColorScaleUtil;
